Add isSaved prop. Refactor react components
