+++ src/registered-stores.js
@@ -1,0 +1,2 @@
+const registeredStores = {};
+export default registeredStores;
